shield/requirements.txt
# SHIELD - Production Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
numpy>=1.24.0
networkx>=3.0
Pillow>=10.0.0

# ML / Embeddings (lightweight + full modes)
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
hdbscan>=0.8.33
scikit-learn>=1.3.0

# Optional: Full mode CV models
# retinaface>=0.0.13
# facenet-pytorch>=2.5.3
# open-clip-torch>=2.20.0

# Utilities
python-multipart>=0.0.6
aiohttp>=3.9.0
